Higher-Order: Implemented new ref handling
Previously, refs on the wrapped component's
instances were created for each branched
component. The refs could then be accessed using
the ComposedComponent's
"getDecoratedComponentInstance" method. However,
this implementation also assigned refs to
Stateless Functional Components (SFCs), which is
not supported by React and leads to console
warnings by React from version 16 and upwards.

The previous behavior has been replaced, removing
the "getDecoratedComponentInstance" method from
the ComposedComponent class, making this a
breaking change.

The new implementation takes a new prop
"decoratedComponentRef" and assigns this to the
"ref" of the wrapped component.