Prevent re-renders when using branch
Before this change, the ComposedComponent returned by branch() would
call setState() in componentDidMount and (if the mapping changed)
componentDidUpdate.

The setState in componentDidMount was mostly unnoticable, but does cause
a small slowdown. The componentDidUpdate behaviour was more broken as
when a ComposedComponent's props changed it would first render its child
components with the new props but the tree state returned by watching
the mapping from the previous props. In componentDidUpdate it would then
correct the mapping and render the child component a second time (this
time correctly) with the new props and tree state returned by watching
the new props.

After this change, we use React's new `calculateDerivedStateFromProps`
to ensure that the child component is rendered once on mount and update;
and that the props are always up to date with the mapping.